+++ vncdotool/rfb.py
@@ -18,12 +18,10 @@ import getpass
 import sys
 import warnings
 import zlib
-from dataclasses import astuple, dataclass
-from struct import Struct, error as StructError, pack, unpack, unpack_from
+from struct import error as StructError, pack, unpack, unpack_from
 from typing import (
     Any,
     Callable,
-    ClassVar,
     Collection,
     Iterator,
     List,
@@ -45,57 +43,11 @@ from twisted.python.failure import Failure
 from . import decoders
 from .const import Encoding, HextileEncoding, AuthTypes, MsgC2S, MsgS2C
 from .keys import Key
+from .pixelformat import PixelFormat
 
-Rect = Tuple[int, int, int, int]
 Ver = Tuple[int, int]
 
 # ~ from twisted.internet import reactor
-
-
-@dataclass(frozen=True)
-class PixelFormat:
-    """:rfc:`6143` §7.4. Pixel Format Data Structure."""
-
-    bpp: int = 32  # u8: bits-per-pixel
-    depth: int = 24  # u8
-    bigendian: bool = False  # u8
-    truecolor: bool = True  # u8
-    redmax: int = 255  # u16
-    greenmax: int = 255  # u16
-    bluemax: int = 255  # u16
-    redshift: int = 0  # u8
-    greenshift: int = 8  # u8
-    blueshift: int = 16  # u8
-
-    STRUCT: ClassVar = Struct("!BB??HHHBBBxxx")
-    VALIDATE: ClassVar = False
-
-    def __post_init__(self) -> None:
-        if not self.VALIDATE:
-            return
-        assert self.bpp in {8, 16, 24, 32}, f"bpp={self.bpp}"
-        assert 1 <= self.depth <= self.bpp, f"depth={self.depth} <= bpp={self.bpp}"
-        if self.truecolor:
-            for max, shift in zip(
-                (self.redmax, self.greenmax, self.bluemax),
-                (self.redshift, self.greenshift, self.blueshift),
-            ):
-                assert 1 <= max <= 0xFFFF, f"1 <= max={max} <= 0xffff"
-                assert max & (max + 1) == 0, f"max={max} not a 2**n-1"
-                assert (
-                    0 <= shift <= self.bpp - max.bit_length()
-                ), f"shift={shift} not in bpp={self.bpp}"
-
-    @property
-    def bypp(self) -> int:  # bytes-per-pixel
-        return (7 + self.bpp) // 8
-
-    @classmethod
-    def from_bytes(cls, block: bytes) -> PixelFormat:
-        return cls(*cls.STRUCT.unpack(block))
-
-    def to_bytes(self) -> bytes:
-        return cast(bytes, self.STRUCT.pack(*astuple(self)))
 
 
 # ZRLE helpers
@@ -429,7 +381,7 @@ class RFBClient(Protocol):  # type: ignore[misc]
 
     def _handleFramebufferUpdate(self, block: bytes) -> None:
         (self.rectangles,) = unpack("!xH", block)
-        self.rectanglePos: list[Rect] = []
+        self.rectanglePos: list[tuple[int, int, int, int]] = []
         self.beginUpdate()
         self._doConnection()
 
@@ -506,7 +458,10 @@ class RFBClient(Protocol):  # type: ignore[misc]
         )
 
     def _finishRectangle(
-        self, decoder: decoders.PixelDecoder, target: decoders.RectBuffer, rect: Rect
+        self,
+        decoder: decoders.PixelDecoder,
+        target: decoders.RectBuffer,
+        rect: tuple[int, int, int, int],
     ) -> None:
         x, y, width, height = rect
         self.updateRectangle(
@@ -536,7 +491,9 @@ class RFBClient(Protocol):  # type: ignore[misc]
         self,
         block: bytes | None,
         generator: Iterator[int],
-        finish: tuple[decoders.PixelDecoder, decoders.RectBuffer, Rect] | None,
+        finish: tuple[
+            decoders.PixelDecoder, decoders.RectBuffer, tuple[int, int, int, int]
+        ] | None,
     ) -> None:
         try:
             size = generator.send(block)
@@ -1145,7 +1102,7 @@ class RFBClient(Protocol):  # type: ignore[misc]
         """called before a series of :meth:`updateRectangle`,
         :meth:`copyRectangle` or :meth:`fillRectangle`."""
 
-    def commitUpdate(self, rectangles: list[Rect] | None = None) -> None:
+    def commitUpdate(self, rectangles: list[tuple[int, int, int, int]] | None = None) -> None:
         """called after a series of :meth:`updateRectangle`, :meth:`copyRectangle`
         or :meth:`fillRectangle` are finished.